% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{assign_sentiment_labels}
\alias{assign_sentiment_labels}
\title{Give labels for scores}
\usage{
assign_sentiment_labels(scores)
}
\arguments{
\item{scores}{Numeric}
}
\description{
Give labels for scores
}
