% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bigrams.R
\name{create_bigrams}
\alias{create_bigrams}
\title{Create bigrams for textual input}
\usage{
create_bigrams(text_df)
}
\arguments{
\item{text_df}{A dataframe}
}
\description{
Create bigrams for textual input
}
