% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{dutch_sentiment_analysis}
\alias{dutch_sentiment_analysis}
\title{Dutch Sentiment Analysis}
\usage{
dutch_sentiment_analysis(text, output = "numeric")
}
\arguments{
\item{text}{A character vector.}

\item{output}{'numeric' (standard) or 'label'.}
}
\value{
Numeric or character, depending on output arg
}
\description{
The sentiment analysis is conducted by using a dutch sentiment dictionary that contains
of +3000 classified words. Pattern matching with the classified words in the dictionary is
performed by computing similarity indices and also takes negation into account ('niet leuk', 'geen charme').
The function scores textual input on a scale from -2 to 2, where negative/positive scores
indicate negative/positive sentiment.
}
\details{
Adjusting the \code{output} arg to 'label' yields textual output ('negative', 'neutral', 'positive')
}
\examples{
text <- c(
 "Ik vond de film matig",
 "De acteurs waren niet slecht",
 "Maar het script was allesbehalve goed",
 "Die kende geen spanning",
 "Zoals je van Tarantino verwacht was het plot fantastisch",
 "Gelukkig waren de bioscoopkaarten goedkoop"
)

# use function
tibble::tibble(
 lines = text,
 scores = dutch_sentiment_analysis(text)
)
}
