% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dutch_sentiment_dictionary}
\alias{dutch_sentiment_dictionary}
\title{A dutch sentiment dictionary}
\format{
A data frame with 5.197 rows and 2 variables:
\describe{
\item{word}{Results of the 5-star rating system, from 1 to 5}
\item{score}{The users opinion of the used product}
}
}
\usage{
dutch_sentiment_dictionary
}
\description{
A dataset containing the dutch sentiment dictionary that includes
+5000 classified words. Word are rated on a scale from -2 to 2, where
negative/positive words indicate negative/positive sentiment.
}
\keyword{datasets}
