% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DutchSentimentAnalysis.R
\docType{package}
\name{DutchSentimentAnalysis}
\alias{DutchSentimentAnalysis}
\title{DutchSentimentAnalysis: A package for performing dutch sentiment analyses.}
\description{
The sentiment analysis is conducted by using a dutch sentiment dictionary that
contains of +5000 classified words. Pattern matching with the classified words
in the dictionary is performed by computing similarity indices and also takes negation
into account ('niet leuk', 'geen charme').
}
\details{
The function scores textual input on a scale from -2 to 2, where negative/positive
scores indicate negative/positive sentiment.
}
\section{Sentiment analysis}{

\code{\link[DutchSentimentAnalysis]{dutch_sentiment_analysis}} \cr
}

\section{Dutch Sentiment Dictionary}{

\code{\link[DutchSentimentAnalysis]{dutch_sentiment_dictionary}} \cr
If you want to lookup a word in the dictionary use:
\code{\link[DutchSentimentAnalysis]{get_word_sentiment}} \cr
}

\section{Validation data}{

\code{\link[DutchSentimentAnalysis]{product_reviews}} \cr
}

