% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_sentiment.R
\name{get_word_sentiment}
\alias{get_word_sentiment}
\title{Lookup the sentimental value of a word in the dictionary}
\usage{
get_word_sentiment(words)
}
\arguments{
\item{words}{Character}
}
\description{
Use this function to get the sentiment of a word in the dutch sentiment dictionary.
}
\examples{
words <- c("goed", "slecht", "lekker")
get_word_sentiment(words)
}
