% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_reviews}
\alias{product_reviews}
\title{Product reviews of earphones.}
\format{
A data frame with 39 rows and 2 variables:
\describe{
\item{Star_rating}{Results of the 5-star rating system, from 1 to 5}
\item{Product_review}{The users opinion of the used product}
}
}
\usage{
product_reviews
}
\description{
A dataset containing product reviews of earphones. Reviewers also
provided their feedback in the form of a 5-star rating.
}
\keyword{datasets}
