% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similarity_in_dictionary.R
\name{search_word}
\alias{search_word}
\title{Search a similar word in the dictionary}
\usage{
search_word(word)
}
\arguments{
\item{word}{Character}
}
\description{
Search a similar word in the dictionary
}
